// types/types.ts
export interface Address {
  street: string;
  street2: string;
  city: string;
  state: string;
  zip: string;
}

export interface Parent {
  _id: string;
  password?: string;
  fullName: string;
  email: string;
  phone: string;
  address:
    | {
        street: string;
        street2: string;
        city: string;
        state: string;
        zip: string;
      }
    | string;
  relationship?: string;
  aauNumber?: string;
  role: string;
  isCoach?: boolean;
  players?: Player[];
  additionalGuardians?: Guardian[];
  createdAt?: string;
  updatedAt?: string;
  isGuardian?: boolean;
  avatar?: string;
  paymentComplete?: boolean;
}
// ...
export interface Guardian {
  id?: string;
  _id?: string;
  fullName: string;
  relationship: string;
  phone: string;
  email: string;
  address:
    | {
        street: string;
        street2: string;
        city: string;
        state: string;
        zip: string;
      }
    | string;
  isCoach: boolean;
  aauNumber: string;
  players?: Player[];
  createdAt?: string;
  updatedAt?: string;
}

export interface Player {
  _id: string;
  parentId?: string;
  fullName: string;
  gender: string;
  grade: string;
  dob: string;
  schoolName: string;
  healthConcerns: string;
  aauNumber: string;
  registrationYear: number;
  season: string;
  createdAt: string;
  updatedAt: string;
  status: string;
  __v: number;
  paymentComplete?: boolean;
}
// ...
export interface TableRecord {
  _id: string;
  fullName: string;
  email: string;
  phone: string;
  address: string | Address;
  role: string;
  type?: 'parent' | 'guardian' | 'coach';
  aauNumber?: string;
  imgSrc?: string;
  createdAt: string;
  status: string;
}